requirements.txt
selenium
pytest
PyRECONSTRUCT
Appium-Python-Client
numpy
theta
setuptools